% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lemma_dictionary.R
\name{make_lemma_dictionary}
\alias{make_lemma_dictionary}
\title{x <- c('the dirtier dog has eaten the pies',
    'that shameful pooch is tricky and sneaky',
    'There are skies of blue and red roses too!'
)
make_lemma_dictionary(x)
make_lemma_dictionary(x, engine = 'treetagger')}
\usage{
make_lemma_dictionary(..., engine = "hunspell", path = NULL)
}
\description{
x <- c('the dirtier dog has eaten the pies',
    'that shameful pooch is tricky and sneaky',
    'There are skies of blue and red roses too!'
)
make_lemma_dictionary(x)
make_lemma_dictionary(x, engine = 'treetagger')
}

