% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstem-package.R
\docType{package}
\name{textstem}
\alias{textstem}
\alias{package-textstem}
\alias{textstem-package}
\title{Tools for Stemming and Lemmatizing Text}
\description{
Tools that stem and lemmatize text.  Stemming is a process that removes
endings such as suffixes.  Lemmatization is the process of grouping 
inflected forms together as a single base form.
}
